sketch: Use toFloat for getting a float value...
Since we now use relative sizes, we need floats instead
of integers for the sizes, otherwise we cannot do stuff
like "half line height".

+++ krita/sketch/Theme.cpp
@@ -407,7 +407,7 @@ void Theme::Private::rebuildFontCache()
             qWarning() << "Could not find font" << map.value("family") << "with style" << map.value("style", "Regular");
 
         float lineHeight = qApp->activeWindow()->height() / lineCount;
-        font.setPixelSize(lineHeight * map.value("size", 1).toInt());
+        font.setPixelSize(lineHeight * map.value("size", 1).toFloat());
 
         fontMap.insert(itr.key(), font);
     }
